requirements.txt
SpeechRecognition
pyttsx3
pyaudio
google-genai
gTTS
pipwin
streamlit


